etl_pipeline/spotipy_client/requirements.txt
spotipy
python-dotenv
pymongo